+++ MyApplication/app/src/main/java/com/example/cds/eattle_prototype_2/model/Tag.java
@@ -4,13 +4,13 @@ package com.example.cds.eattle_prototype_2.model;
  * Created by GA on 2015. 3. 31..
  */
 public class Tag {
-    long id;
+    int id;
     String name;
 
     public Tag() {
     }
 
-    public Tag(long id, String name) {
+    public Tag(int id, String name) {
         this.id = id;
         this.name = name;
     }
@@ -19,11 +19,11 @@ public class Tag {
         this.name = name;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
